+++ NEWS.md
@@ -1,3 +1,5 @@
 # CBTF 0.1.0 (2024-11-14)
 
 - First version of the package.
+- The package website is now available at
+  https://mcol.github.io/caught-by-the-fuzz/