+++ NEWS.md
@@ -1,3 +1,5 @@
+# Current development version
+
 # CBTF 0.1.0 (2024-11-14)
 
 - First version of the package.