+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Return an object with the raw results from `fuzz()` and provide S3 methods
   for `summary()` and `print()` (#1).
+- Skip functions that call `readline()`, as in an interactive session they
+  stall waiting for user input (#4).
 
 # CBTF 0.1.0 (2024-11-14)